move build_metric_meta to metrics API and move hidden operations to constant
