+++ src/sentry/sentry_metrics/querying/metadata/utils.py
@@ -1,6 +1,8 @@
 from sentry.snuba.metrics import get_mri
 from sentry.snuba.metrics.naming_layer.mri import is_mri
 
+METRICS_API_HIDDEN_OPERATIONS = ["p50", "p75", "p90", "p95", "p99"]
+
 
 def convert_metric_names_to_mris(metric_names: list[str]) -> list[str]:
     mris: list[str] = []
